Extract response parser
